Fix data loading (#2007)
