--- packages/client/src/app/components/modals/kamiBridge/index.ts
@@ -1,1 +1,0 @@
-export { KamiBridge } from './KamiBridge';
