+++ packages/client/src/app/components/modals/kamiPortal/index.ts
@@ -1,0 +1,1 @@
+export { KamiPortalModal } from './KamiPortal';
