+++ packages/client/src/app/components/modals/naming/EmaBoard.tsx
@@ -21,8 +21,8 @@ import { Stage } from './Stage';
 
 const REFRESH_INTERVAL = 2000;
 
-export const EmaBoard: UIComponent = {
-  id: 'EmaBoard',
+export const EmaBoardModal: UIComponent = {
+  id: 'EmaBoardModal',
   requirement: (layers) =>
     interval(1000).pipe(
       map(() => {
@@ -45,118 +45,119 @@ export const EmaBoard: UIComponent = {
                 live: 2,
                 progress: 3600,
               }),
-            getItemBalance: (inventory: Inventory[], index: number) => getInventoryBalance(inventory, index),
+            getItemBalance: (inventory: Inventory[], index: number) =>
+              getInventoryBalance(inventory, index),
             getItem: (index: number) => getItemByIndex(world, components, index),
           },
         };
       })
     ),
   Render: ({ network, data, utils }) => {
-      const { accountEntity, spender } = data;
-      const { actions, api } = network;
-      const { getAccount, getItem, getKamis } = utils;
-      const { selectedAddress, apis: ownerAPIs } = useNetwork();
-      const { balances: tokenBals } = useTokens();
-      const { modals } = useVisibility();
+    const { accountEntity, spender } = data;
+    const { actions, api } = network;
+    const { getAccount, getItem, getKamis } = utils;
+    const { selectedAddress, apis: ownerAPIs } = useNetwork();
+    const { balances: tokenBals } = useTokens();
+    const { modals } = useVisibility();
 
-      const [tick, setTick] = useState(Date.now());
-      const [kamis, setKamis] = useState<Kami[]>([]);
-      const [account, setAccount] = useState<Account>(NullAccount);
-      const [selected, setSelected] = useState<Kami>(NullKami);
-      const [holyDustItem, setHolyDustItem] = useState<Item>(NullItem);
-      const [onyxItem, setOnyxItem] = useState<Item>(NullItem);
-      const [onyxInfo, setOnyxInfo] = useState<BalPair>({ allowance: 0, balance: 0 });
+    const [tick, setTick] = useState(Date.now());
+    const [kamis, setKamis] = useState<Kami[]>([]);
+    const [account, setAccount] = useState<Account>(NullAccount);
+    const [selected, setSelected] = useState<Kami>(NullKami);
+    const [holyDustItem, setHolyDustItem] = useState<Item>(NullItem);
+    const [onyxItem, setOnyxItem] = useState<Item>(NullItem);
+    const [onyxInfo, setOnyxInfo] = useState<BalPair>({ allowance: 0, balance: 0 });
 
-      /////////////////
-      // SUBSCRIPTIONS
+    /////////////////
+    // SUBSCRIPTIONS
 
-      useEffect(() => {
-        setOnyxItem(getItem(ONYX_INDEX));
-        setHolyDustItem(getItem(HOLY_DUST_INDEX));
+    useEffect(() => {
+      setOnyxItem(getItem(ONYX_INDEX));
+      setHolyDustItem(getItem(HOLY_DUST_INDEX));
 
-        const refreshClock = () => setTick(Date.now());
-        const timerId = setInterval(refreshClock, REFRESH_INTERVAL);
-        return () => clearInterval(timerId);
-      }, []);
+      const refreshClock = () => setTick(Date.now());
+      const timerId = setInterval(refreshClock, REFRESH_INTERVAL);
+      return () => clearInterval(timerId);
+    }, []);
 
-      useEffect(() => {
-        if (!modals.emaBoard) return;
-        setAccount(getAccount());
-        setKamis(getKamis());
-      }, [modals.emaBoard, accountEntity, tick]);
+    useEffect(() => {
+      if (!modals.emaBoard) return;
+      setAccount(getAccount());
+      setKamis(getKamis());
+    }, [modals.emaBoard, accountEntity, tick]);
 
-      useEffect(() => {
-        const onyxInfo = tokenBals.get(onyxItem.address!);
-        setOnyxInfo(onyxInfo ?? { allowance: 0, balance: 0 });
-      }, [onyxItem, spender, tick]);
+    useEffect(() => {
+      const onyxInfo = tokenBals.get(onyxItem.address!);
+      setOnyxInfo(onyxInfo ?? { allowance: 0, balance: 0 });
+    }, [onyxItem, spender, tick]);
 
-      /////////////////
-      // ACTIONS
+    /////////////////
+    // ACTIONS
 
-      // approve the spend of an ERC20 token
-      const approveOnyxTx = (amt: number) => {
-        const api = ownerAPIs.get(selectedAddress);
-        if (!api) return console.error(`API not established for ${selectedAddress}`);
+    // approve the spend of an ERC20 token
+    const approveOnyxTx = (amt: number) => {
+      const api = ownerAPIs.get(selectedAddress);
+      if (!api) return console.error(`API not established for ${selectedAddress}`);
 
-        const actionID = uuid() as EntityID;
-        actions.add({
-          id: actionID,
-          action: 'Approve token',
-          params: [onyxItem.address, spender, amt],
-          description: `Approve ${amt} ${onyxItem.name} to be spent`,
-          execute: async () => {
-            return api.erc20.approve(onyxItem.address!, spender, amt);
-          },
-        });
-        return actionID;
-      };
+      const actionID = uuid() as EntityID;
+      actions.add({
+        id: actionID,
+        action: 'Approve token',
+        params: [onyxItem.address, spender, amt],
+        description: `Approve ${amt} ${onyxItem.name} to be spent`,
+        execute: async () => {
+          return api.erc20.approve(onyxItem.address!, spender, amt);
+        },
+      });
+      return actionID;
+    };
 
-      const renameTx = (kami: Kami, name: string) => {
-        const actionID = uuid() as EntityID;
-        actions.add({
-          id: actionID,
-          action: 'KamiName',
-          params: [kami.id, name],
-          description: `Renaming ${kami.name} to ${name}`,
-          execute: async () => {
-            return api.player.pet.name(kami.id, name);
-          },
-        });
-        return actionID;
-      };
+    const renameTx = (kami: Kami, name: string) => {
+      const actionID = uuid() as EntityID;
+      actions.add({
+        id: actionID,
+        action: 'KamiName',
+        params: [kami.id, name],
+        description: `Renaming ${kami.name} to ${name}`,
+        execute: async () => {
+          return api.player.pet.name(kami.id, name);
+        },
+      });
+      return actionID;
+    };
 
-      const onyxRenameTx = (kami: Kami, name: string) => {
-        const api = ownerAPIs.get(selectedAddress);
-        if (!api) return console.error(`API not established for ${selectedAddress}`);
+    const onyxRenameTx = (kami: Kami, name: string) => {
+      const api = ownerAPIs.get(selectedAddress);
+      if (!api) return console.error(`API not established for ${selectedAddress}`);
 
-        const actionID = uuid() as EntityID;
-        actions.add({
-          action: 'KamiOnyxRename',
-          params: [kami.id, name],
-          description: `Renaming ${kami.name} to ${name} with ONYX`,
-          execute: async () => {
-            return api.pet.onyx.rename(kami.id, name);
-          },
-        });
-        return actionID;
-      };
+      const actionID = uuid() as EntityID;
+      actions.add({
+        action: 'KamiOnyxRename',
+        params: [kami.id, name],
+        description: `Renaming ${kami.name} to ${name} with ONYX`,
+        execute: async () => {
+          return api.pet.onyx.rename(kami.id, name);
+        },
+      });
+      return actionID;
+    };
 
-      return (
-        <ModalWrapper
-          id='emaBoard'
-          header={<ModalHeader title='Ema Board' icon={KamiIcon} />}
-          canExit
-          noPadding
-          truncate
-        >
-          <Stage
-            actions={{ onyxApprove: approveOnyxTx, rename: renameTx, onyxRename: onyxRenameTx }}
-            data={{ account, kami: selected, onyxItem, holyDustItem, onyxInfo }}
-            state={{ tick }}
-            utils={utils}
-          />
-          <Carousel kamis={kamis} state={{ selected, setSelected }} />
-        </ModalWrapper>
-      );
+    return (
+      <ModalWrapper
+        id='emaBoard'
+        header={<ModalHeader title='Ema Board' icon={KamiIcon} />}
+        canExit
+        noPadding
+        truncate
+      >
+        <Stage
+          actions={{ onyxApprove: approveOnyxTx, rename: renameTx, onyxRename: onyxRenameTx }}
+          data={{ account, kami: selected, onyxItem, holyDustItem, onyxInfo }}
+          state={{ tick }}
+          utils={utils}
+        />
+        <Carousel kamis={kamis} state={{ selected, setSelected }} />
+      </ModalWrapper>
+    );
   },
 };
