+++ packages/client/src/app/components/modals/obol/index.ts
@@ -1,0 +1,1 @@
+export { ObolModal } from './Obol';
