+++ packages/client/src/app/components/modals/reveal/Reveal.tsx
@@ -12,8 +12,8 @@ import { queryDTCommits } from 'network/shapes/Droptable';
 import { useWatchBlockNumber } from 'wagmi';
 import { Commits } from './Commits';
 
-export const Reveal: UIComponent = {
-  id: 'Reveal',
+export const RevealModal: UIComponent = {
+  id: 'RevealModal',
   requirement: (layers) =>
     interval(1000).pipe(
       map(() => {
@@ -29,92 +29,92 @@ export const Reveal: UIComponent = {
       })
     ),
   Render: ({ network, data }) => {
-      const { commits } = data;
-      const {
-        actions,
-        api,
-        components: { State },
-        world,
-        localSystems: { DTRevealer },
-      } = network;
+    const { commits } = data;
+    const {
+      actions,
+      api,
+      components: { State },
+      world,
+      localSystems: { DTRevealer },
+    } = network;
 
-      const [blockNumber, setBlockNumber] = useState(BigInt(0));
+    const [blockNumber, setBlockNumber] = useState(BigInt(0));
 
-      useWatchBlockNumber({
-        onBlockNumber: (n) => {
-          setBlockNumber(n);
+    useWatchBlockNumber({
+      onBlockNumber: (n) => {
+        setBlockNumber(n);
+      },
+    });
+
+    useEffect(() => {
+      commits.map((commit) => DTRevealer.add(commit));
+    }, [commits, blockNumber]);
+
+    useEffect(() => {
+      execute();
+    }, [blockNumber]);
+
+    /////////////////
+    // REVEAL LOGIC
+
+    async function execute() {
+      const commits = DTRevealer.extractQueue();
+      if (commits.length === 0) return;
+
+      const actionIndex = await revealTx(commits);
+      DTRevealer.finishReveal(actionIndex, commits);
+    }
+
+    async function overrideExecute(commits: EntityID[]) {
+      if (commits.length === 0) return;
+
+      DTRevealer.forceQueue(commits);
+      const actionIndex = await revealTx(commits);
+      DTRevealer.finishReveal(actionIndex, commits);
+    }
+
+    /////////////////
+    // TRANSACTIONS
+
+    const revealTx = async (commits: EntityID[]): Promise<EntityIndex> => {
+      const actionIndex = actions.add({
+        action: 'Droptable reveal',
+        params: [commits],
+        description: `Inspecting item contents`,
+        execute: async () => {
+          return api.player.droptable.reveal(commits);
         },
       });
+      await waitForActionCompletion(actions.Action, actionIndex);
+      return actionIndex;
+    };
 
-      useEffect(() => {
-        commits.map((commit) => DTRevealer.add(commit));
-      }, [commits, blockNumber]);
+    /////////////////
+    // UTILS
 
-      useEffect(() => {
-        execute();
-      }, [blockNumber]);
+    const getCommitState = (id: EntityID): string => {
+      const entity = world.entityToIndex.get(id);
+      if (!entity) return 'EXPIRED';
+      const state = getComponentValue(State, entity)?.value as string;
+      return state ?? 'EXPIRED';
+    };
 
-      /////////////////
-      // REVEAL LOGIC
-
-      async function execute() {
-        const commits = DTRevealer.extractQueue();
-        if (commits.length === 0) return;
-
-        const actionIndex = await revealTx(commits);
-        DTRevealer.finishReveal(actionIndex, commits);
-      }
-
-      async function overrideExecute(commits: EntityID[]) {
-        if (commits.length === 0) return;
-
-        DTRevealer.forceQueue(commits);
-        const actionIndex = await revealTx(commits);
-        DTRevealer.finishReveal(actionIndex, commits);
-      }
-
-      /////////////////
-      // TRANSACTIONS
-
-      const revealTx = async (commits: EntityID[]): Promise<EntityIndex> => {
-        const actionIndex = actions.add({
-          action: 'Droptable reveal',
-          params: [commits],
-          description: `Inspecting item contents`,
-          execute: async () => {
-            return api.player.droptable.reveal(commits);
-          },
-        });
-        await waitForActionCompletion(actions.Action, actionIndex);
-        return actionIndex;
-      };
-
-      /////////////////
-      // UTILS
-
-      const getCommitState = (id: EntityID): string => {
-        const entity = world.entityToIndex.get(id);
-        if (!entity) return 'EXPIRED';
-        const state = getComponentValue(State, entity)?.value as string;
-        return state ?? 'EXPIRED';
-      };
-
-      return (
-        <ModalWrapper
-          id='reveal'
-          header={<ModalHeader title='Commits' icon={SettingsIcon} />}
-          overlay
-          canExit
-        >
-          <Container>
-            <Commits
-              data={{ commits: commits, blockNumber: Number(blockNumber) }}
-              actions={{ revealTx: overrideExecute }}
-              utils={{ getCommitState }}
-            />
-          </Container>
-        </ModalWrapper>
-      );
+    return (
+      <ModalWrapper
+        id='reveal'
+        header={<ModalHeader title='Commits' icon={SettingsIcon} />}
+        overlay
+        canExit
+      >
+        <Container>
+          <Commits
+            data={{ commits: commits, blockNumber: Number(blockNumber) }}
+            actions={{ revealTx: overrideExecute }}
+            utils={{ getCommitState }}
+          />
+        </Container>
+      </ModalWrapper>
+    );
   },
 };
