+++ packages/client/src/app/components/modals/reveal/index.ts
@@ -1,1 +1,1 @@
-export { Reveal } from './Reveal';
+export { RevealModal } from './Reveal';
