+++ packages/client/src/network/shapes/Kami/types.ts
@@ -104,7 +104,7 @@ export const get = (
   if (options?.progress) kami.progress = getProgress(comps, entity);
   if (options?.rerolls) kami.rerolls = getRerolls(comps, entity);
   if (options?.skills) kami.skills = getSkills(world, comps, entity);
-  if (options?.stats) kami.stats = getStats(world, comps, entity);
+  if (options?.stats) kami.stats = getStats(world, comps, entity, true);
   if (options?.time) kami.time = getTimes(comps, entity);
   if (options?.traits) kami.traits = getTraits(world, comps, entity);
